requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.9.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
discord.py==2.3.2
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.18.0
frozenlist==1.4.1
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.7
jiter==0.9.0
mcp==1.6.0
multidict==6.0.5
mysql-connector-python==8.2.0
openai==1.30.1
packaging==24.2
protobuf==4.21.12
pydantic==2.7.2
pydantic-settings==2.8.1
pydantic_core==2.18.3
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.5
sse-starlette==2.2.1
starlette==0.46.1
tokenizers==0.21.1
tqdm==4.66.4
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
yarl==1.9.4
